refactor(service/): 인터페이스와 구현체 이름 변경
프레임워크에 의존적인 이름을 사용하지 않기 위헤 인터페이스과 구현체 이름
변경

BREAKING CHANGE: 인터페이스, 구현체 이름  변경
as-is:
interface:
    LineService
    StationService
concrete class:
    SpringLineServic
    SpringStationService

to-be:
interface:
    CommonLineService
    CommonStationService
concrete class:
    LineServic
    StationService
