+++ src/main/java/wooteco/subway/service/LineService.java
@@ -1,18 +1,49 @@
 package wooteco.subway.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.dto.LineRequest;
 
-public interface LineService {
+@Service
+public class LineService implements CommonLineService {
 
-    Line save(final LineRequest lineRequest);
+    private final LineDao lineDao;
 
-    Line findById(final Long id);
+    public LineService(final LineDao lineDao) {
+        this.lineDao = lineDao;
+    }
 
-    List<Line> findAll();
+    @Override
+    @Transactional
+    public Line save(final LineRequest lineRequest) {
+        final Line line = new Line(lineRequest.getName(), lineRequest.getColor());
+        return lineDao.save(line);
+    }
 
-    void update(final Long id, final LineRequest lineRequest);
+    @Override
+    @Transactional(readOnly = true)
+    public Line findById(final Long id) {
+        return lineDao.findById(id);
+    }
 
-    void deleteById(final Long id);
+    @Override
+    @Transactional(readOnly = true)
+    public List<Line> findAll() {
+        return lineDao.findAll();
+    }
+
+    @Override
+    @Transactional
+    public void update(final Long id, final LineRequest lineRequest) {
+        lineDao.update(id, lineRequest);
+    }
+
+    @Override
+    @Transactional
+    public void deleteById(final Long id) {
+        lineDao.deleteById(id);
+    }
 }
