+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,14 +1,39 @@
 package wooteco.subway.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 
-public interface StationService {
+@Service
+public class StationService implements CommonStationService {
 
-    Station save(final StationRequest stationRequest);
+    private final StationDao stationDao;
 
-    List<Station> findAll();
+    public StationService(final StationDao stationDao) {
+        this.stationDao = stationDao;
+    }
 
-    void deleteById(final Long id);
+    @Transactional
+    @Override
+    public Station save(final StationRequest stationRequest) {
+        final Station station = new Station(stationRequest.getName());
+        return stationDao.save(station);
+    }
+
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<Station> findAll() {
+        return stationDao.findAll();
+    }
+
+
+    @Transactional
+    @Override
+    public void deleteById(final Long id) {
+        stationDao.deleteById(id);
+    }
 }
